# Improve loggers performance