--- a/internal/slogadapter/gokit.go
+++ b/internal/slogadapter/gokit.go
@@ -19,25 +19,58 @@ type slogAdapter struct {
 	h slog.Handler
 }
 
+var (
+	_ log.Logger = (*slogAdapter)(nil)
+)
+
+// Enabled implements logging.EnabledAware interface.
+func (sa slogAdapter) Enabled(ctx context.Context, l slog.Level) bool {
+	return sa.h.Enabled(ctx, l)
+}
+
 func (sa slogAdapter) Log(kvps ...interface{}) error {
-	// We don't know what the log level or message are yet, so we set them to
-	// defaults until we iterate through kvps.
-	//
+	// Find the log level first, starting with the default.
+	recordLevel := slog.LevelInfo
+	for i := 0; i < len(kvps); i += 2 {
+		if kvps[i] == level.Key() {
+			if i+1 < len(kvps) {
+				value := kvps[i+1]
+				levelValue, _ := value.(level.Value)
+				switch levelValue {
+				case level.DebugValue():
+					recordLevel = slog.LevelDebug
+				case level.InfoValue():
+					recordLevel = slog.LevelInfo
+				case level.WarnValue():
+					recordLevel = slog.LevelWarn
+				case level.ErrorValue():
+					recordLevel = slog.LevelError
+				}
+				break
+			}
+		}
+	}
+
+	// Do not build the record if the level is not enabled.
+	if !sa.h.Enabled(context.Background(), recordLevel) {
+		return nil
+	}
+
 	// Since there's a pattern of wrapping loggers at different depths in go-kit,
 	// we can't consistently know who the caller is, so we set no value for pc
 	// here.
-	rec := slog.NewRecord(time.Now(), slog.LevelInfo, "", 0)
+	rec := slog.NewRecord(time.Now(), recordLevel, "", 0)
 
 	for i := 0; i < len(kvps); i += 2 {
 		var key, value any
 
 		if i+1 < len(kvps) {
-			key = kvps[i+0]
+			key = kvps[i]
 			value = kvps[i+1]
 		} else {
 			// Mismatched pair
 			key = "!BADKEY"
-			value = kvps[i+0]
+			value = kvps[i]
 		}
 
 		if key == "msg" || key == "message" {
@@ -46,19 +79,7 @@ func (sa slogAdapter) Log(kvps ...interface{}) error {
 		}
 
 		if key == level.Key() {
-			levelValue, _ := value.(level.Value)
-
-			switch levelValue {
-			case level.DebugValue():
-				rec.Level = slog.LevelDebug
-			case level.InfoValue():
-				rec.Level = slog.LevelInfo
-			case level.WarnValue():
-				rec.Level = slog.LevelWarn
-			case level.ErrorValue():
-				rec.Level = slog.LevelError
-			}
-
+			// Already handled
 			continue
 		}
 
@@ -68,8 +89,5 @@ func (sa slogAdapter) Log(kvps ...interface{}) error {
 		})
 	}
 
-	if !sa.h.Enabled(context.Background(), rec.Level) {
-		return nil
-	}
 	return sa.h.Handle(context.Background(), rec)
 }