<!--

CONTRIBUTORS GUIDE: https://github.com/grafana/agent/blob/main/docs/developer/contributing.md#updating-the-changelog

If this is your first PR or you have not contributed in a while, we recommend
taking the time to review the guide. It gives helpful instructions for
contributors around things like how to update the changelog.

-->

#### PR Description

It's recommended too review the first two commits of this PR separately:
* [First one](https://github.com/grafana/agent/pull/5452/commits/c9d638b3917ecac138afe53c60339100842a00ce) adds benchmarks and includes the results in a comment
* [Second one](https://github.com/grafana/agent/pull/5452/commits/3ed08cfd5598a7d3f336f7af529083d314b5db62) adds a few optimisations for the logging and includes the new benchmark results in a comment.

Summary of optimisations:
* Look for the log level first in log adapter to avoid costly allocations if the record will not be logged
* Implement our own `level` package that can replace gokit's `level.Debug(logger)...` - our `level` package is level-aware and will be checking whether given level is enabled in the underlying logger (if it can, opportunistically. The idea is that we can replace the import with our own `level` and should get performance boost. 

In the long run we should still move to `slog` as it's more performant and simpler.

#### Which issue(s) this PR fixes

<!-- Uncomment the following line if you want that GitHub issue gets automatically closed after merging the PR -->
<!-- Fixes #issue_id -->

#### Notes to the Reviewer

#### PR Checklist

<!-- Remove items that do not apply. For completed items, change [ ] to [x]. -->

- [ ] CHANGELOG.md updated
- [ ] Documentation added
- [x] Tests updated
- [ ] Config converters updated