@thampiotr 2023-10-12
Here's a `benchstat` comparison - thanks @rfratto for tool suggestion
```
benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/grafana/agent/pkg/flow/logging
                                   │   old.txt    │               new.txt               │
                                   │    sec/op    │   sec/op     vs base                │
Logging_NoLevel_Prints-8              1.552µ ± 1%   1.542µ ± 0%   -0.64% (p=0.002 n=10)
Logging_NoLevel_Drops-8              334.20n ± 2%   47.18n ± 2%  -85.88% (p=0.000 n=10)
Logging_GoKitLevel_Drops_Sprintf-8    484.1n ± 0%   376.7n ± 5%  -22.19% (p=0.000 n=10)
Logging_GoKitLevel_Drops-8            503.8n ± 0%   201.2n ± 1%  -60.06% (p=0.000 n=10)
Logging_GoKitLevel_Prints-8           1.735µ ± 0%   1.741µ ± 0%   +0.35% (p=0.003 n=10)
Logging_Slog_Drops-8                  50.51n ± 1%   37.04n ± 0%  -26.67% (p=0.000 n=10)
Logging_Slog_Prints-8                 1.194µ ± 1%   1.169µ ± 1%   -2.01% (p=0.000 n=10)
Logging_FlowLevel_Drops-8                           74.12n ± 1%
Logging_FlowLevel_Prints-8                          1.740µ ± 0%
geomean                               539.1n        335.3n       -39.00%

                                   │  old.txt   │               new.txt                │
                                   │    B/op    │    B/op     vs base                  │
Logging_NoLevel_Prints-8             408.0 ± 0%   392.0 ± 0%   -3.92% (p=0.000 n=10)
Logging_NoLevel_Drops-8              64.00 ± 0%   16.00 ± 0%  -75.00% (p=0.000 n=10)
Logging_GoKitLevel_Drops_Sprintf-8   464.0 ± 0%   352.0 ± 0%  -24.14% (p=0.000 n=10)
Logging_GoKitLevel_Drops-8           528.0 ± 0%   480.0 ± 0%   -9.09% (p=0.000 n=10)
Logging_GoKitLevel_Prints-8          873.0 ± 0%   873.0 ± 0%        ~ (p=1.000 n=10) ¹
Logging_Slog_Drops-8                 32.00 ± 0%   16.00 ± 0%  -50.00% (p=0.000 n=10)
Logging_Slog_Prints-8                64.00 ± 0%   47.00 ± 0%  -26.56% (p=0.000 n=10)
Logging_FlowLevel_Drops-8                         176.0 ± 0%
Logging_FlowLevel_Prints-8                        873.0 ± 0%
geomean                              196.8        168.1       -32.97%
¹ all samples are equal

                                   │   old.txt   │               new.txt                │
                                   │  allocs/op  │ allocs/op   vs base                  │
Logging_NoLevel_Prints-8              13.00 ± 0%   12.00 ± 0%   -7.69% (p=0.000 n=10)
Logging_NoLevel_Drops-8               7.000 ± 0%   1.000 ± 0%  -85.71% (p=0.000 n=10)
Logging_GoKitLevel_Drops_Sprintf-8   11.000 ± 0%   9.000 ± 0%  -18.18% (p=0.000 n=10)
Logging_GoKitLevel_Drops-8           11.000 ± 0%   6.000 ± 0%  -45.45% (p=0.000 n=10)
Logging_GoKitLevel_Prints-8           17.00 ± 0%   17.00 ± 0%        ~ (p=1.000 n=10) ¹
Logging_Slog_Drops-8                  2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
Logging_Slog_Prints-8                 5.000 ± 0%   3.000 ± 0%  -40.00% (p=0.000 n=10)
Logging_FlowLevel_Drops-8                          3.000 ± 0%
Logging_FlowLevel_Prints-8                         17.00 ± 0%
geomean                               7.871        4.919       -43.82%
¹ all samples are equal
```